Force delete any old presubmit branches before patching
We expect the code tree to be reused between tests (and in fact,
within the same test by different target/config matrix options).  Make
sure old presubmit branches are deleted before patching.

This change also adds an option to bypass this behavior because we may
want to run two patches in a row (e.g. to patch two dependent sets of
CLs into the same tree.)

Change-Id: I05b0f52875ab41a0f897902c21541383eb6bc7bc