[gitsync] Mirror only heads and tags to github, not changes
Doing a default mirror clone and push will try to sync all of the
refs/* space from gerrit to github, including refs/changes/* which
may contain things that aren't very useful (or possible) to mirror.
This changes the sync script to explicitly pull down only heads and
tags and mirror those.

Fixes IN-77

Change-Id: Ifa5ca2d2fc69b7c29fb5fa941a98c62de505d19d